+++ examples/example1/example.py
@@ -18,6 +18,10 @@ class HelloWorldService(MicroService):
     host = "127.0.0.1"
     port = 5000
 
+    accessible_at = "http://localhost:5000/api"
+
+    service_registry_urls = ["http://localhost:8000/api"]
+
     extra_handlers = [
         (r"/app", IndexHandler),
     ]
@@ -60,4 +64,4 @@ class HelloWorldService(MicroService):
 
 if __name__ == '__main__':
     service = HelloWorldService()
-    service.start()
+    service.get_cli()()